refactor: person details to clean architecture

+++ domain/src/main/java/br/dev/singular/overview/domain/repository/Generics.kt
@@ -25,7 +25,7 @@ interface GetPage<T, P> {
 }
 
 interface GetById<T> {
-    suspend fun getById(id: Int): T?
+    suspend fun getById(id: Long): T?
 }
 
 data class Page<T>(
